fixed mobile issue
