+++ propertydating/src/components/pages/swipePage/ImageArrowLeft.js
@@ -16,7 +16,6 @@ export default function ImageArrowLeft({ setCurrentImage, currentImage }) {
           viewBox="0 0 20 20"
           fill="currentColor"
           onClick={handleClick}
-          onTouchStart={handleClick}
         >
           <path
             fillRule="evenodd"
