+++ propertydating/src/css/vanilla-css/nav.css
@@ -27,6 +27,10 @@
   text-align: center;
   text-transform: uppercase;
 }
+.top-nav__logo img{
+  max-height: 40px;
+  object-fit: contain;
+}
 .top-nav__login{
   grid-area: login;
   display: flex;
